+++ RInAppMessaging/Classes/Models/Impression.swift
@@ -1,5 +1,5 @@
 /// Model for an impression object.
 internal struct Impression: Encodable {
     let type: ImpressionType
-    let timestamp: Int
+    let timestamp: Int64
 }
